参考以太坊(https://zhuanlan.zhihu.com/p/28830859 and https://www.jianshu.com/p/9e56faac2437)实现了难度调整算法。

+++ src/main/kotlin/nebular/core/BlockHeader.kt
@@ -4,5 +4,5 @@ import org.joda.time.DateTime
 import java.math.BigInteger
 
 class BlockHeader(val version: Int, val height: Long, val parentHash: ByteArray,
-    val coinBase: ByteArray, val time: DateTime, val difficulty: Int,
+    val coinBase: ByteArray, val time: DateTime, val difficulty: Long,
     val nonce: Int, val totalDifficulty: BigInteger, val stateRoot: ByteArray, val trxTrieRoot: ByteArray)
